add byte in FlipTriangle

+++ triangulation/src/main/java/triangulation/FliperDelaunay.java
@@ -17,7 +17,7 @@ public class FliperDelaunay implements Fliper {
 
     @Override
     public void add(TriangleStructure triangle, int index) {
-        buffer.add(new FlipStructure(triangle, index));
+        buffer.add(new FlipStructure(triangle,(byte) index));
     }
 
     @Override
